Add ability to load plugins from jars
This commit adds a second lein project, which is a very simple
web service like the ones in the main project.  Using this
second project you can run `lein jar` to build a jar version
of a plugin for the service framework, drop it into the
plugins directory of the main project, and fire everything up.
The web service should be successfully loaded as a plugin from
the jar.